TERRAWIZE/expertosy_web/backend/requirements.txt
flask==2.3.2
flask-cors==4.0.0
openai==1.3.0
python-dotenv==1.0.0
openpyxl==3.1.2
asyncio==3.4.3
gunicorn==20.1.0 